Atualização da leitura do arquivo.

+++ estrutura.h
@@ -3,8 +3,8 @@
 
 typedef struct
 {
-    long *jogoInicial;    // Array com 9 longs, cada um representando uma linha do jogo
-    char *jogoPreenchido; // Array com 91 chars, representando os 81 números, as quebras de linha e o \0
+    int *jogoInicial;    // Array com 9 ints, cada um representando uma linha do jogo
+    char *jogoPreenchido; // Array com 90 chars, representando os 81 números, as quebras de linha e o \0
 } tJogo;
 
 typedef struct
@@ -22,6 +22,6 @@ extern void apagarLista(tListaIndexada lista);
 extern int compararJogo(const void *a, const void *b);
 extern void ordenarLista(tListaIndexada lista);
 
-extern char *buscaInterpolacao(tListaIndexada lista, long *jogo, int *qOperacoes);
+extern char *buscaInterpolacao(tListaIndexada lista, int *jogo, int *qOperacoes);
 
 #endif
